% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{BCS}
\alias{BCS}
\title{Compute the between class scatter matrix}
\usage{
BCS(splitclusters)
}
\arguments{
\item{splitclusters}{A list of dataframes (from the output of split()) with 
scaled data from each cluster}
}
\value{
returns the between class scatter matrix
}
\description{
Takes in a list of dataframes with scaled data (output from split()) and 
returns the between class scatter matrix
}
\examples{
data(iris)
iris_input <- dplyr::select(iris, -c("Species"))
sp_dfs <- split(iris_input, f = iris$Species)
bcs_mat <- iDA::BCS(sp_dfs)

}
