% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{betweenclass_scatter_matrix}
\alias{betweenclass_scatter_matrix}
\title{Compute the between class scatter matrix}
\usage{
betweenclass_scatter_matrix(splitclusters)
}
\arguments{
\item{splitclusters}{A list of dataframes (from the output of split_clusters) with scaled data from each cluster}
}
\value{
returns the between class scatter matrix
}
\description{
Takes in a list of dataframes with scaled data (output from split_clusters) and returns the between class scatter matrix
}
