% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ida.R
\name{ChangePoint.RUD}
\alias{ChangePoint.RUD}
\title{Changepoint of normalized coverage}
\usage{
ChangePoint.RUD(GR, SampleInfo, cv)
}
\arguments{
\item{GR}{a genomic range of unioned 3'utr frame}

\item{SampleInfo}{a data frame with columns named sample,Bam,Seq_depth,Group}

\item{cv}{a data frame generate by GenerateReadsPerBase function}
}
\value{
A data frame of candidate poly(A) sites with Cov_L and Cov_S
}
\description{
Changepoint of normalized coverage
}
\details{
function which compute the change points of reads coverage of 3'utr frame
}
\author{
Xia Xiao
}
