% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{decompose}
\alias{decompose}
\title{Find singular vectors for Sw^-1 * Sb}
\usage{
decompose(withinclust_sc_mat, betweenclust_sc_mat, num_clusts = 10)
}
\arguments{
\item{withinclust_sc_mat}{The within class scatter matrix (output from withinclust_sc_mat())}

\item{betweenclust_sc_mat}{The between class scatter matrix (output from betweenclust_sc_mat())}

\item{num_clusts}{The number of clusters}
}
\value{
returns the between class scatter matrix
}
\description{
Takes in a within cluster scatter matrix, a between cluster scatter matrix, and returns the singular value decomposition for Rayleigh quotient between them
}
