% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ida.R
\name{DetectPAs}
\alias{DetectPAs}
\title{Changepoint of normalized coverage}
\usage{
DetectPAs(GR, SampleInfo, cut.off = 30, ...)
}
\arguments{
\item{GR}{a genomic range of unioned 3'utr frame}

\item{SampleInfo}{a data frame with columns named sample,Bam,Seq_depth,Group}
}
\value{
A data frame of candidate poly(A) sites
}
\description{
Changepoint of normalized coverage
}
\details{
function which compute the change points of reads coverage of 3'utr frame
}
\author{
Xia Xiao
}
