% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ida.R
\name{GenerateReadsPerBase}
\alias{GenerateReadsPerBase}
\title{Generate reads per base of genomic range}
\usage{
GenerateReadsPerBase(GR, SampleInfo, mapq = 255, ...)
}
\arguments{
\item{GR}{a genomic range of unioned 3'utr frame}

\item{SampleInfo}{a data frame with columns named sample,Bam,Seq_depth,Group}

\item{mapq}{a numeric for mapping quality,default is 255 for uniquelly mapped reads of STAR output,
Can be set to 30 when using Bowtie}
}
\value{
A data frame of reads per base of given samples
}
\description{
Generate reads per base of genomic range
}
\details{
function which generate reads per base,default for properly paired ,uniquelly mapped reads
and normalized by sequence depth
}
\author{
Xia Xiao
}
