% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{getFeatureWeights}
\alias{getFeatureWeights}
\title{Get Feature Weights from iDA}
\usage{
getFeatureWeights(object)
}
\arguments{
\item{object}{(SE or SCE) an SE or SCE object in which to fill in feature weights}
}
\value{
a data.frame of features x LDs containing the feature weights from iDA
}
\description{
Get Feature Weights from iDA
}
