% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,DESeqDataSet-method}
\alias{iDA,DESeqDataSet-method}
\title{Set method for DESeqDataSet to input data to iDA}
\usage{
\S4method{iDA}{DESeqDataSet}(object, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{...}{Additional arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Set method for DESeqDataSet to input data to iDA
}
\examples{
countData <- matrix(1:100,ncol=4)
condition <- factor(c("A","A","B","B"))
dds <- DESeq2::DESeqDataSetFromMatrix(countData, S4Vectors::DataFrame(condition), ~ condition)
set.seed(11)
dds <- iDA(dds)

}
