% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,matrix-method}
\alias{iDA,matrix-method}
\title{Set method for matrix input to iDA}
\usage{
\S4method{iDA}{matrix}(object, nFeatures = 2000, ...)
}
\arguments{
\item{object}{The normalized matrix (feature x sample) to run iDA on}

\item{nFeatures}{The number of HVG features to use in reduction}

\item{...}{Additional arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights as a list
}
\description{
Set method for matrix input to iDA
}
\examples{
exp <- matrix(rpois(20000, 5), ncol=20)
colnames(exp) <- paste0("donor", seq_len(ncol(exp)))
rownames(exp) <- paste0("gene", seq_len(nrow(exp)))
logexp <- logexp <- log2(exp + 1)
set.seed(11)
se <- iDA(logexp)

}
