% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,Seurat-method}
\alias{iDA,Seurat-method}
\title{Method for Seurat object to input data to iDA}
\usage{
\S4method{iDA}{Seurat}(object, assay, ...)
}
\arguments{
\item{object}{The single cell experiment object to run iDA on}

\item{assay}{The assay to take counts from}

\item{...}{Additional arguments passed to object constructors}
}
\value{
Seurat object with iDA cell weights and gene weights stored in object[["iDA"]] and cluster assignments stored in rowLabels
}
\description{
Method for Seurat object to input data to iDA
}
