% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,SingleCellExperiment-method}
\alias{iDA,SingleCellExperiment-method}
\title{Method for SingleCellExperiment object to input data to iDA}
\usage{
\S4method{iDA}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{The single cell experiment object to run iDA on}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
SingleCellExperiment object with iDA cell weights and gene weights stored in reducedDims and cluster assignemts
stored in rowLabels
}
\description{
Method for SingleCellExperiment object to input data to iDA
}
