% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,SummarizedExperiment-method}
\alias{iDA,SummarizedExperiment-method}
\title{Set method for SummarizedExperiment to input data to iDA}
\usage{
\S4method{iDA}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{...}{Additional arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Set method for SummarizedExperiment to input data to iDA
}
\examples{
data(airway, package="airway")
se <- airway
set.seed(11)
se <- iDA(se)

}
