% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA}
\alias{iDA}
\alias{iDA,matrix-method}
\alias{iDA,SummarizedExperiment-method}
\alias{iDA,DESeqDataSet-method}
\alias{iDA,SingleCellExperiment-method}
\title{Generic method to input data to iDA}
\usage{
iDA(object, ...)

\S4method{iDA}{matrix}(object, nFeatures = 2000, ...)

\S4method{iDA}{SummarizedExperiment}(object, nFeatures = 2000, ...)

\S4method{iDA}{DESeqDataSet}(object, nFeatures = 2000, ...)

\S4method{iDA}{SingleCellExperiment}(object, nFeatures = 2000, ...)
}
\arguments{
\item{object}{The single cell experiment object to run iDA on}

\item{...}{Additional arguments passed to object constructors}

\item{nFeatures}{The number of HVG features to use in reduction}
}
\value{
iDA output with clustering, gene weights, and cell weights

iDA output with clustering, gene weights, and cell weights as a list

iDA output with clustering, gene weights, and cell weights

iDA output with clustering, gene weights, and cell weights

SingleCellExperiment object with iDA cell weights and gene weights 
stored in reducedDims and cluster assignments stored in rowLabels
}
\description{
Generic method to input data to iDA

Set method for matrix input to iDA

Set method for SummarizedExperiment to input data to iDA

Set method for DESeqDataSet to input data to iDA

Method for SingleCellExperiment object to input data to iDA
}
\examples{

exp <- matrix(rpois(20000, 5), ncol=20)
colnames(exp) <- paste0("donor", seq_len(ncol(exp)))
rownames(exp) <- paste0("gene", seq_len(nrow(exp)))
logexp <- logexp <- log2(exp + 1)
set.seed(11)
ida <- iDA(logexp)

exp <- matrix(rpois(20000, 5), ncol=20)
colnames(exp) <- paste0("donor", seq_len(ncol(exp)))
rownames(exp) <- paste0("gene", seq_len(nrow(exp)))
logexp <- logexp <- log2(exp + 1)
conditions <- factor(rep(1:4, 5))
se <- SummarizedExperiment::SummarizedExperiment(
assays = list(counts = exp, logcounts = logexp),
colData = data.frame(conditions = conditions))
set.seed(11)
se <- iDA(se)

library(SingleCellExperiment)
data(sc_sample_data, package = "scPipe")
sce <- SingleCellExperiment(assays = list(counts = as.matrix(sc_sample_data)))
set.seed(11)
sce <- iDA(sce)

}
