% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{setFeatureWeights}
\alias{setFeatureWeights}
\title{Set Feature Weights from iDA}
\usage{
setFeatureWeights(object, feature.weights)
}
\arguments{
\item{object}{(SE or SCE) an SE or SCE object in which to fill in feature weights}

\item{feature.weights}{(matrix) matrix of features x LDs which is output from iDA}
}
\value{
an SE or SCE object with rowData() for the iDA feature weights
}
\description{
Set Feature Weights from iDA
}
