% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{VariableGenes}
\alias{VariableGenes}
\title{Find Variable Genes}
\usage{
VariableGenes(data.use, dispersion.cutoff, mean.low.cutoff, mean.high.cutoff)
}
\arguments{
\item{data.use}{(data.frame) features are rows, samples are columns. Rows must be named.}

\item{dispersion.cutoff}{(numeric) rows returned will have scaled dispersion higher provided cutoff}

\item{mean.low.cutoff}{(numeric) rows returned will have average higher than this cutoff}

\item{mean.high.cutoff}{(numeric) rows returned will have average lower than this cutoff}
}
\value{
(character) a list of row names with high dispersion rows
}
\description{
Identify rows in a data frame with high sclaed dispersion. This rule was
taken from the Seurat package.
}
