% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{VariableGenesGeneric}
\alias{VariableGenesGeneric}
\title{Find Variable Genes}
\usage{
VariableGenesGeneric(
  NormCounts,
  dispersion.cutoff = 1,
  mean.low.cutoff = 0.1,
  mean.high.cutoff = 8
)
}
\arguments{
\item{NormCounts}{(data.frame) features are rows, samples are columns. 
Rows must be named.}

\item{dispersion.cutoff}{(numeric) rows returned will have scaled dispersion 
higher provided cutoff}

\item{mean.low.cutoff}{(numeric) rows returned will have average higher than 
this cutoff}

\item{mean.high.cutoff}{(numeric) rows returned will have average lower than
this cutoff}
}
\value{
(character) a list of row names with high dispersion rows
}
\description{
Identify rows in a data frame with high sclaed dispersion. This rule was
taken from the Seurat package.
}
\examples{
data("sc_sample_data")
sce <- SingleCellExperiment(list(counts = as.matrix(sc_sample_data)))
logcounts(sce) <- normalizeCounts(sce,  size.factors = sizeFactors(sce))
var.features <- iDA::VariableGenesGeneric(logcounts(sce))

}
