% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{WCS}
\alias{WCS}
\title{Compute each cluster's within class scatter matrix}
\usage{
WCS(splitclusters, diag = FALSE)
}
\arguments{
\item{splitclusters}{A list of dataframes with scaled data from each cluster 
(output from split_clusters())}

\item{diag}{if off diagonal entries in within class scatter matrix should be
zeroed}
}
\value{
returns the within class scatter matrix
}
\description{
Takes in the output from split_clusters() and computes the within class 
scatter matrix
}
