% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{withinclass_scattermatrix_QDA}
\alias{withinclass_scattermatrix_QDA}
\title{Compute each cluster's within class scatter matrix (for QDA)}
\usage{
withinclass_scattermatrix_QDA(splitclusters, diag = FALSE)
}
\arguments{
\item{splitclusters}{A list of dataframes with scaled data from each cluster (output from split_clusters())}

\item{diag}{if off diagonal entries in within class scatter matrix should be zeroed}
}
\value{
returns the within class scatter matrix
}
\description{
Takes in the output from split_clusters() and computes the within class scatter matrix
}
