% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{decomposeSVD}
\alias{decomposeSVD}
\title{Compute LDA using within and between cluster covariance matrices}
\usage{
decomposeSVD(WCSmat, BCSmat, nu = 10)
}
\arguments{
\item{WCSmat}{Output from WCS (matrix) of within class scatter}

\item{BCSmat}{Output from BCS (matrix) of between class scatter}

\item{nu}{The number of columns (eigenvectors) to keep}
}
\value{
returns the between class scatter matrix
}
\description{
Takes in within and between cluster covariance matrices (output from WCS
and BSC) and returns the eigenvectors and eigenvalues for the decomp.
}
