% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{getLouvain}
\alias{getLouvain}
\title{Cluster Determination}
\usage{
getLouvain(SNN)
}
\arguments{
\item{SNN}{a matrix of shared nearest neighbors (output from getSNN)}
}
\value{
a list of identities for clustering
}
\description{
Identify clusters of cells by a shared nearest neighbor (SNN) modularity
optimization based clustering algorithm. Optimize the modularity function to
determine clusters. For a full description of the algorithms, see Waltman and
van Eck (2013) \emph{The European Physical Journal B}.
}
