% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{getLouvain}
\alias{getLouvain}
\title{Louvain Cluster Determination}
\usage{
getLouvain(data.use, k.param, prune.SNN)
}
\arguments{
\item{data.use}{(matrix) Matrix with scaled data to find nearest neighbors}

\item{k.param}{(numeric) Defines k for the k-nearest neighbor algorithm}

\item{prune.SNN}{(numeric) Sets the cutoff for acceptable Jaccard index when
computing the neighborhood overlap for the SNN construction. Any edges with
values less than or equal to this will be set to 0 and removed from the SNN
graph. Essentially sets the strigency of pruning (0 --- no pruning, 1 ---
prune everything).}
}
\value{
a list of identities for clustering
}
\description{
Identify clusters of cells by a shared nearest neighbor (SNN) modularity
optimization based clustering algorithm. Optimize the modularity function to
determine clusters. For a full description of the algorithms, see Waltman and
van Eck (2013) \emph{The European Physical Journal B}.
}
