% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{getLouvain}
\alias{getLouvain}
\title{Cluster Determination}
\usage{
getLouvain(SNN, set.seed = set.seed)
}
\arguments{
\item{SNN}{a matrix of shared nearest neighbors (output from getSNN)}

\item{set.seed}{Seed of the random number generator.}
}
\value{
a list of identities for clustering
}
\description{
Identify clusters of cells by a shared nearest neighbor (SNN) modularity
optimization based clustering algorithm. Optimize the modularity function to
determine clusters. For a full description of the algorithms, see Waltman and
van Eck (2013) \emph{The European Physical Journal B}.
}
\examples{
data("sc_sample_data")
sce <- SingleCellExperiment(assays = list(counts = as.matrix(sc_sample_data)))
logcounts(sce) <- normalizeCounts(sce,  size.factors = sizeFactors(sce))
snn <- iDA::getSNN(data.use = logcounts(sce))
clusters <- iDA::getLouvain(snn, set.seed = 11)

}
