% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,DESeqDataSet-method}
\alias{iDA,DESeqDataSet-method}
\title{Set method for DESeqDataSet to input data to iDA}
\usage{
\S4method{iDA}{DESeqDataSet}(object, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{...}{Additional arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Set method for DESeqDataSet to input data to iDA
}
\examples{
data(airway, package="airway")
se <- airway
dds <- DESeq2::DESeqDataSet(se, ~ dex)
set.seed(11)
dds <- iDA(dds)

}
