% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,matrix-method}
\alias{iDA,matrix-method}
\title{Set method for matrix to input data to iDA}
\usage{
\S4method{iDA}{matrix}(object, scaled, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{scaled}{Is the input matrix normalized and scaled already?}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Set method for matrix to input data to iDA
}
