% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\name{iDA,SummarizedExperiment-method}
\alias{iDA,SummarizedExperiment-method}
\title{Set method for SummarizedExperiment to input data to iDA}
\usage{
\S4method{iDA}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{...}{Additional arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Set method for SummarizedExperiment to input data to iDA
}
\examples{
exp <- matrix(rpois(20000, 5), ncol=20)
colnames(exp) <- paste0("donor", seq_len(ncol(exp)))
rownames(exp) <- paste0("gene", seq_len(nrow(exp)))
logexp <- logexp <- log2(exp + 1)
conditions <- factor(rep(1:4, 5))
se <- SummarizedExperiment::SummarizedExperiment(
assays = list(counts = exp, logcounts = logexp),
colData = data.frame(conditions = conditions))
set.seed(11)
se <- iDA(se)

}
