% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{split_clusters}
\alias{split_clusters}
\title{Split dataset by cluster}
\usage{
split_clusters(data, clusterIDcol)
}
\arguments{
\item{data}{A dataframe to be split}

\item{clusterIDcol}{The column  in data with cluster identifiers}
}
\value{
n number of dataframes for each cluster's data
}
\description{
Takes in dataframe to be split and an cluster identifier column and outputs n number of dataframes (n = number of clusters)
}
